refactor(chat): extract ConversationView from ChatPanel
ChatPanel becomes a thin shell that handles loading/error/permission
banner branches. The actual conversation render lives in
ConversationView. Same behavior, smaller files. useChatTurn signature
gains threadId so ConversationView can pass it through (currently null;
wired in B7).
